// src/model/interfaces/viatics/GActivities.tsx
import { GExpenses } from './GExpenses';
export class GActivities {
    BookingCode: number;
    Budget: number;
    Currency: string;
    DateEnd: Date;
    DateLegalize: Date |  null;
    DateSta: Date;
    DateSync: Date;
    Description: string;
    EditedLegalize: boolean;
    ExpenseGroupCash: number;
    IDEntity: number;
    IDGroup: string;
    IDParent: string;
    IDUser: number;
    IDUserLegalize: number;
    LegalizeCash: number;
    LegalizeNote: string;
    LegalizeState: string;
    Note: string;
    State: string;
    TotalLegalize: number;
    Edited?: boolean;
    PigeOnHole?: boolean;
    visibleReadOnly?: boolean;
    TypeOwner: string;
    UserName: string;
    EntityName: string;
    ApproverUserName: string;
    createBySystem: string;
    Trm: number;
    DeliveryTrm: number;
    maxDateLegalize: Date | null;
    countExpenses: number;
    totalExpense: number;
    HasAttach: boolean;
    RefundDate: Date;
    TotalRefund: number;
    Requestdate: Date;
    ClosedType?: string;
    interNal: string;
    items: GExpenses[];
    entregadoUsd: boolean;
    tipoConsultaReporteAnticipos: string;
    fechaEntregaAnticipo: Date;
    IDApproverUser: number;
    EmailRequester: string;
    UserApprover: number;
    SecondUserApprover: number;
    SecondApprover: string;
    viat_fecAprobacionParcial: string;
    IdUserParcialLegalize: 0;
    DateParcialLegalize: Date;
  }
  