// src/model/classes/viatics/Establishment.tsx
export class Establishment {
    ID: number;
    IDEntity: number;
    Name: string;
    Description: string;
    IDEstablisment: string;
    IDUser: number;
    Datetime: Date;
    ERPCode: string;
    HomologationCode: string;
    AuxCode1: string;
    AuxCode2: string;
    Status: string;
    Address: string;
    Phone: string;

}